refactor(app): extract composer input mapping
