+++ src/app.rs
@@ -16,6 +16,7 @@ pub mod chat_store;
 pub mod community_bridge;
 pub mod community_hierarchy;
 pub mod composer;
+pub mod composer_input_mapping;
 pub mod contact_avatars;
 pub mod download_worker;
 pub mod events;
